==> data_workflow_notebooks/3_preprocessing.ipynb <==
import json
import os

import polars as pl

data_path = "data"
tabular_data_path = os.path.join(data_path, "tabular")
vocabulary_path = os.path.join(data_path, "vocabulary.json")

# splits = ["train", "validation", "test"]
splits = ["train_img_only", "test_img_only"]
with open(vocabulary_path) as f:
    vocabulary = json.load(f)

multi_hot = [pl.col("description").str.split(" ").list.contains(word).cast(pl.Int8).alias(word) for word in vocabulary]

for split in splits:
    dataset_path = os.path.join(tabular_data_path, f"{split}.csv")
    processed_data_path = os.path.join(tabular_data_path, f"{split}_processed.csv")
    dataset = pl.read_csv(dataset_path)
    if "description" in dataset.columns:
        dataset = dataset.with_columns(multi_hot).drop("description").select(pl.exclude("class"), pl.col("class")).select(pl.exclude("file_name"), pl.col("file_name"))
    dataset.write_csv(processed_data_path)
